refactor: remove unnessesary class
